requirements.txt
diff-match-patch==20121119
microsofttranslator==0.8
requests==2.11.1
six==1.10.0
translate-toolkit==1.13.0
